query dsl

+++ src/main/java/com/excilys/computerDatabase/page/Page.java
@@ -19,16 +19,16 @@ public class Page {
 
 	private int limit = 10;
 	private static int actualPage = 1;
-	private int nbComputer = 0;
-	private String search="";
+	private long nbComputer = 0;
+	private String search = "";
 	private String order = "name";
 	private boolean asc = true;
 	private ComputerService computerService;
-	private int nbPage;
+	private long nbPage;
 	private int offset;
 	@Autowired
 	private ComputerMapper computerMapper;
-	
+
 	@Autowired
 	public Page(ComputerService computerService) {
 		this.computerService = computerService;
@@ -44,10 +44,10 @@ public class Page {
 	}
 
 	public void setLimitString(String limitString) throws BadEntriException {
-		if(limitString != null ) {
+		if (limitString != null) {
 			try {
 				limit = Integer.parseInt(limitString);
-			} catch(NumberFormatException e) {
+			} catch (NumberFormatException e) {
 				throw new BadEntriException("the limit is not an integer");
 			}
 			updateNbPage();
@@ -56,31 +56,31 @@ public class Page {
 	}
 
 	private void updateNbPage() {
-		if(limit == 0 ) {
+		if (limit == 0) {
 			nbPage = 1;
 		} else {
 			nbPage = (nbComputer + limit - 1) / limit;
 		}
 	}
-	
+
 	public void updateNbComputer() {
 		nbComputer = computerService.count(search);
 		updateNbPage();
 	}
-	
+
 	private void updateOffset() {
 		offset = (actualPage - 1) * limit;
 	}
 
 	public void setSearch(String search) {
-		if(search != null) {
+		if (search != null) {
 			this.search = search;
 			updateNbComputer();
 		}
 	}
 
 	public void setOrder(String order) {
-		if(order != null) {
+		if (order != null) {
 			this.order = order;
 			asc = !asc;
 		}
@@ -91,25 +91,25 @@ public class Page {
 	}
 
 	public List<ComputerDto> getComputersDto() {
-		List<Computer> computers = computerService.getPage(limit, offset,search, order, asc);
+		List<Computer> computers = computerService.getPage(limit, offset, search, order, asc);
 		return computers.stream().map(computer -> computerMapper.computerToDto(computer)).collect(Collectors.toList());
 	}
-	
+
 	public List<Computer> getComputers() {
-		return computerService.getPage(limit, offset,search, order, asc);
+		return computerService.getPage(limit, offset, search, order, asc);
 	}
 
 	public Object getNbPage() {
 		return nbPage;
 	}
 
-	public int getNbComputer() {
+	public long getNbComputer() {
 		return nbComputer;
 	}
 
 	public void setActualPageString(String newPageString) throws BadEntriException {
 		int newPage;
-		if(newPageString != null) {
+		if (newPageString != null) {
 			try {
 				newPage = Integer.parseInt(newPageString);
 			} catch (NumberFormatException e) {
@@ -120,9 +120,9 @@ public class Page {
 			updateOffset();
 		}
 	}
-	
+
 	private void checkPage(int newPage) throws BadEntriException {
-		if(newPage < 1 || newPage > nbPage) {
+		if (newPage < 1 || newPage > nbPage) {
 			throw new BadEntriException("page not correct");
 		}
 	}
@@ -146,14 +146,14 @@ public class Page {
 			}
 		}
 	}
-	
-	public Map<String,Object> getInfo() {
-		Map<String,Object> map = new HashMap<>();
+
+	public Map<String, Object> getInfo() {
+		Map<String, Object> map = new HashMap<>();
 		map.put("computers", getComputers());
 		map.put("nbPage", getNbPage());
 		map.put("page", getActualPage());
 		map.put("nbComputer", getNbComputer());
 		return map;
 	}
-	
+
 }
